Renamed views to pages
