--- client/src/views/VisualiserDashboard.module.scss
@@ -1,6 +1,0 @@
-// .container {
-//     // position: static;
-//     width: 100vw;
-//     height: 100vh;
-//     // height: calc(100vh - 64px);
-// }
